+++ src/main/java/com/example/coconote/api/chatRoom/controller/ChatRoomController.java
@@ -23,6 +23,15 @@ public class ChatRoomController {
         return "/chat/room";
     }
 
+    // 채팅방 입장 화면
+    @GetMapping("/room/enter/{roomId}")
+    public String roomDetail(Model model, @PathVariable String roomId) {
+        model.addAttribute("roomId", roomId);
+        return "/chat/roomdetail";
+    }
+
+
+
     @GetMapping("/rooms")
     @ResponseBody
     public List<ChatRoom> room() {
@@ -33,13 +42,6 @@ public class ChatRoomController {
     @ResponseBody
     public ChatRoom createRoom(@RequestParam String name) {
         return chatRoomRepository.createChatRoom(name);
-    }
-
-    // 채팅방 입장 화면
-    @GetMapping("/room/enter/{roomId}")
-    public String roomDetail(Model model, @PathVariable String roomId) {
-        model.addAttribute("roomId", roomId);
-        return "/chat/roomdetail";
     }
