--- src/main/java/com/example/coconote/api/chatRoom/controller/WebSocketController.java
@@ -1,25 +1,0 @@
-package com.example.coconote.api.chatRoom.controller;
-
-// import 생략...
-
-import com.example.coconote.api.chatRoom.dto.ChatMessage;
-import com.example.coconote.api.chatRoom.entity.MessageType;
-import lombok.RequiredArgsConstructor;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.simp.SimpMessageSendingOperations;
-import org.springframework.stereotype.Controller;
-
-@RequiredArgsConstructor
-@Controller
-public class WebSocketController {
-
-    private final SimpMessageSendingOperations messagingTemplate;
-
-    @MessageMapping("/chat/message")
-    public void message(ChatMessage message) {
-        if (MessageType.ENTER.equals(message.getType()))
-            message.setMessage(message.getSender() + "님이 입장하셨습니다.");
-
-        messagingTemplate.convertAndSend("/sub/chat/room/" + message.getRoomId(), message);
-    }
-}
