+++ src/main/java/com/example/coconote/api/thread/dto/requset/ChatMessage2.java
@@ -1,10 +1,12 @@
-package com.example.coconote.api.chatRoom.dto;
+package com.example.coconote.api.thread.dto.requset;
 
 import com.example.coconote.api.chatRoom.entity.MessageType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.util.List;
 
 @Data
 @Builder
@@ -14,5 +16,7 @@ public class ChatMessage2 {
     private MessageType type;
     private Long channelId;
     private Long sender;
+    private Long parentThreadId;
     private String message;
+    private List<String> files;
 }
