+++ src/main/java/com/example/coconote/api/thread/entity/Thread.java
@@ -44,7 +44,7 @@ public class Thread extends BaseEntity {
         return ThreadResDto.builder()
                 .id(this.id)
                 .memberName(this.member.getNickname())
-                .createdTime(this.getCreatedTime())
+                .createdTime(this.getCreatedTime().toString())
                 .content(this.content)
                 .files(this.files)
                 .tags(tags)
@@ -56,7 +56,7 @@ public class Thread extends BaseEntity {
         return ThreadResDto.builder()
                 .id(this.id)
                 .memberName(this.member.getNickname())
-                .createdTime(this.getCreatedTime())
+                .createdTime(this.getCreatedTime().toString())
                 .content(this.content)
                 .files(this.files)
                 .childThreads(childThreadList)
