+++ src/main/java/com/example/coconote/api/thread/service/ThreadService.java
@@ -30,7 +30,7 @@ public class ThreadService {
 
     public ThreadResDto createThread(ThreadCreateReqDto dto) {
         //TODO: jwt토큰이 완성되면 memberId 는 불러오면됨
-        Member member = memberRepository.findById(dto.getMemberId()).orElseThrow(()-> new EntityNotFoundException("해당멤버가 없습니다."));
+        Member member = memberRepository.findById(dto.getSenderId()).orElseThrow(()-> new EntityNotFoundException("해당멤버가 없습니다."));
         Thread parentThread = threadRepository.findById(dto.getParentId()).orElse(null);
         Channel channel = channelRepository.findById(dto.getChannelId()).orElseThrow(()->new EntityNotFoundException("해당 채널이 없습니다."));
         Thread thread = threadRepository.save(dto.toEntity(member,parentThread, channel));
